requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cssbeautifier==1.14.6
Deprecated==1.2.13
Django==4.1.1
django-csp==3.7
django-defender==0.9.5
django-recaptcha==3.0.0
django-taggit==3.0.0
djlint==1.12.0
docopt==0.6.2
EditorConfig==0.12.3
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.3
importlib-metadata==4.12.0
jsbeautifier==1.14.6
packaging==21.3
pathspec==0.9.0
Pillow==9.2.0
pipreqs==0.4.11
pyparsing==3.0.9
python-decouple==3.6
PyYAML==6.0
redis==4.3.4
regex==2022.8.17
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
tzdata==2022.2
urllib3==1.26.12
wrapt==1.14.1
yarg==0.1.9
zipp==3.8.1
